refactor: changing things around to support this as a common sdk
The existing js-client-sdk only works for the web but the actual
differences required to support react-native are very small. I'm
refactoring the client sdk into a common package that is shared
by both javascript client SDKs. This requires small changes to
how the SDK works.

Basically storage is different between the two and the native SDK
has no concept of session storage. This reflects that.